+++ src/storage/src/mem_table.rs
@@ -12,38 +12,23 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::cmp::Ordering;
 use std::collections::btree_map::Entry;
 use std::collections::BTreeMap;
-use std::future::Future;
-use std::ops::Bound::{Excluded, Included, Unbounded};
+use std::ops::Bound::{Included, Unbounded};
 use std::ops::RangeBounds;
-use std::sync::Arc;
 
 use bytes::Bytes;
-use futures::{pin_mut, Stream, StreamExt};
-use futures_async_stream::try_stream;
-use itertools::Itertools;
-use risingwave_common::bitmap::Bitmap;
-use risingwave_common::catalog::{TableId, TableOption};
-use risingwave_common::hash::{VirtualNode, VnodeBitmapExt};
 use risingwave_common_estimate_size::{EstimateSize, KvSize};
-use risingwave_hummock_sdk::key::{prefixed_range_with_vnode, FullKey, TableKey, TableKeyRange};
-use risingwave_hummock_sdk::table_watermark::WatermarkDirection;
+use risingwave_hummock_sdk::key::TableKey;
 use thiserror::Error;
 use thiserror_ext::AsReport;
 use tracing::error;
 
-use crate::error::{StorageError, StorageResult};
 use crate::hummock::iterator::{Backward, Forward, FromRustIterator, RustIteratorBuilder};
 use crate::hummock::shared_buffer::shared_buffer_batch::{SharedBufferBatch, SharedBufferBatchId};
-use crate::hummock::utils::{
-    do_delete_sanity_check, do_insert_sanity_check, do_update_sanity_check, sanity_check_enabled,
-};
+use crate::hummock::utils::sanity_check_enabled;
 use crate::hummock::value::HummockValue;
-use crate::memory::{RangeKv, RangeKvStateStore};
 use crate::row_serde::value_serde::ValueRowSerde;
-use crate::storage_value::StorageValue;
 use crate::store::*;
 pub type ImmutableMemtable = SharedBufferBatch;
 
@@ -425,343 +410,6 @@ impl KeyOp {
                 format!("Update({:?}, {:?})", &before, &after)
             }
         }
-    }
-}
-
-#[try_stream(ok = StateStoreKeyedRow, error = StorageError)]
-pub(crate) async fn merge_stream<'a>(
-    mem_table_iter: impl Iterator<Item = (&'a TableKey<Bytes>, &'a KeyOp)> + 'a,
-    inner_stream: impl Stream<Item = StorageResult<StateStoreKeyedRow>> + 'static,
-    table_id: TableId,
-    epoch: u64,
-    rev: bool,
-) {
-    let inner_stream = inner_stream.peekable();
-    pin_mut!(inner_stream);
-
-    let mut mem_table_iter = mem_table_iter.fuse().peekable();
-
-    loop {
-        match (inner_stream.as_mut().peek().await, mem_table_iter.peek()) {
-            (None, None) => break,
-            // The mem table side has come to an end, return data from the shared storage.
-            (Some(_), None) => {
-                let (key, value) = inner_stream.next().await.unwrap()?;
-                yield (key, value)
-            }
-            // The stream side has come to an end, return data from the mem table.
-            (None, Some(_)) => {
-                let (key, key_op) = mem_table_iter.next().unwrap();
-                match key_op {
-                    KeyOp::Insert(value) | KeyOp::Update((_, value)) => {
-                        yield (FullKey::new(table_id, key.clone(), epoch), value.clone())
-                    }
-                    _ => {}
-                }
-            }
-            (Some(Ok((inner_key, _))), Some((mem_table_key, _))) => {
-                debug_assert_eq!(inner_key.user_key.table_id, table_id);
-                let mut ret = inner_key.user_key.table_key.cmp(mem_table_key);
-                if rev {
-                    ret = ret.reverse();
-                }
-                match ret {
-                    Ordering::Less => {
-                        // yield data from storage
-                        let (key, value) = inner_stream.next().await.unwrap()?;
-                        yield (key, value);
-                    }
-                    Ordering::Equal => {
-                        // both memtable and storage contain the key, so we advance both
-                        // iterators and return the data in memory.
-
-                        let (_, key_op) = mem_table_iter.next().unwrap();
-                        let (key, old_value_in_inner) = inner_stream.next().await.unwrap()?;
-                        match key_op {
-                            KeyOp::Insert(value) => {
-                                yield (key.clone(), value.clone());
-                            }
-                            KeyOp::Delete(_) => {}
-                            KeyOp::Update((old_value, new_value)) => {
-                                debug_assert!(old_value == &old_value_in_inner);
-
-                                yield (key, new_value.clone());
-                            }
-                        }
-                    }
-                    Ordering::Greater => {
-                        // yield data from mem table
-                        let (key, key_op) = mem_table_iter.next().unwrap();
-
-                        match key_op {
-                            KeyOp::Insert(value) => {
-                                yield (FullKey::new(table_id, key.clone(), epoch), value.clone());
-                            }
-                            KeyOp::Delete(_) => {}
-                            KeyOp::Update(_) => unreachable!(
-                                "memtable update should always be paired with a storage key"
-                            ),
-                        }
-                    }
-                }
-            }
-            (Some(Err(_)), Some(_)) => {
-                // Throw the error.
-                return Err(inner_stream.next().await.unwrap().unwrap_err());
-            }
-        }
-    }
-}
-
-pub struct RangeKvLocalStateStore<R: RangeKv> {
-    mem_table: MemTable,
-    inner: RangeKvStateStore<R>,
-
-    epoch: Option<u64>,
-
-    table_id: TableId,
-    op_consistency_level: OpConsistencyLevel,
-    table_option: TableOption,
-    vnodes: Arc<Bitmap>,
-}
-
-impl<R: RangeKv> RangeKvLocalStateStore<R> {
-    pub fn new(inner: RangeKvStateStore<R>, option: NewLocalOptions) -> Self {
-        Self {
-            inner,
-            mem_table: MemTable::new(option.op_consistency_level.clone()),
-            epoch: None,
-            table_id: option.table_id,
-            op_consistency_level: option.op_consistency_level,
-            table_option: option.table_option,
-            vnodes: option.vnodes,
-        }
-    }
-}
-
-impl<R: RangeKv> LocalStateStore for RangeKvLocalStateStore<R> {
-    type Iter<'a> = impl StateStoreIter + 'a;
-    type RevIter<'a> = impl StateStoreIter + 'a;
-
-    async fn get(
-        &self,
-        key: TableKey<Bytes>,
-        read_options: ReadOptions,
-    ) -> StorageResult<Option<Bytes>> {
-        match self.mem_table.buffer.get(&key) {
-            None => self.inner.get(key, self.epoch(), read_options).await,
-            Some(op) => match op {
-                KeyOp::Insert(value) | KeyOp::Update((_, value)) => Ok(Some(value.clone())),
-                KeyOp::Delete(_) => Ok(None),
-            },
-        }
-    }
-
-    #[allow(clippy::manual_async_fn)]
-    fn iter(
-        &self,
-        key_range: TableKeyRange,
-        read_options: ReadOptions,
-    ) -> impl Future<Output = StorageResult<Self::Iter<'_>>> + Send + '_ {
-        async move {
-            let iter = self
-                .inner
-                .iter(key_range.clone(), self.epoch(), read_options)
-                .await?;
-            Ok(FromStreamStateStoreIter::new(Box::pin(merge_stream(
-                self.mem_table.iter(key_range),
-                iter.into_stream(to_owned_item),
-                self.table_id,
-                self.epoch(),
-                false,
-            ))))
-        }
-    }
-
-    #[allow(clippy::manual_async_fn)]
-    fn rev_iter(
-        &self,
-        key_range: TableKeyRange,
-        read_options: ReadOptions,
-    ) -> impl Future<Output = StorageResult<Self::RevIter<'_>>> + Send + '_ {
-        async move {
-            let iter = self
-                .inner
-                .rev_iter(key_range.clone(), self.epoch(), read_options)
-                .await?;
-            Ok(FromStreamStateStoreIter::new(Box::pin(merge_stream(
-                self.mem_table.rev_iter(key_range),
-                iter.into_stream(to_owned_item),
-                self.table_id,
-                self.epoch(),
-                true,
-            ))))
-        }
-    }
-
-    fn insert(
-        &mut self,
-        key: TableKey<Bytes>,
-        new_val: Bytes,
-        old_val: Option<Bytes>,
-    ) -> StorageResult<()> {
-        match old_val {
-            None => self.mem_table.insert(key, new_val)?,
-            Some(old_val) => self.mem_table.update(key, old_val, new_val)?,
-        };
-        Ok(())
-    }
-
-    fn delete(&mut self, key: TableKey<Bytes>, old_val: Bytes) -> StorageResult<()> {
-        Ok(self.mem_table.delete(key, old_val)?)
-    }
-
-    async fn flush(&mut self) -> StorageResult<usize> {
-        let buffer = self.mem_table.drain().into_parts();
-        let mut kv_pairs = Vec::with_capacity(buffer.len());
-        for (key, key_op) in buffer {
-            match key_op {
-                // Currently, some executors do not strictly comply with these semantics. As
-                // a workaround you may call disable the check by initializing the
-                // state store with `op_consistency_level=Inconsistent`.
-                KeyOp::Insert(value) => {
-                    if sanity_check_enabled() {
-                        do_insert_sanity_check(
-                            &key,
-                            &value,
-                            &self.inner,
-                            self.epoch(),
-                            self.table_id,
-                            self.table_option,
-                            &self.op_consistency_level,
-                        )
-                        .await?;
-                    }
-                    kv_pairs.push((key, StorageValue::new_put(value)));
-                }
-                KeyOp::Delete(old_value) => {
-                    if sanity_check_enabled() {
-                        do_delete_sanity_check(
-                            &key,
-                            &old_value,
-                            &self.inner,
-                            self.epoch(),
-                            self.table_id,
-                            self.table_option,
-                            &self.op_consistency_level,
-                        )
-                        .await?;
-                    }
-                    kv_pairs.push((key, StorageValue::new_delete()));
-                }
-                KeyOp::Update((old_value, new_value)) => {
-                    if sanity_check_enabled() {
-                        do_update_sanity_check(
-                            &key,
-                            &old_value,
-                            &new_value,
-                            &self.inner,
-                            self.epoch(),
-                            self.table_id,
-                            self.table_option,
-                            &self.op_consistency_level,
-                        )
-                        .await?;
-                    }
-                    kv_pairs.push((key, StorageValue::new_put(new_value)));
-                }
-            }
-        }
-        self.inner.ingest_batch(
-            kv_pairs,
-            vec![],
-            WriteOptions {
-                epoch: self.epoch(),
-                table_id: self.table_id,
-            },
-        )
-    }
-
-    fn epoch(&self) -> u64 {
-        self.epoch.expect("should have set the epoch")
-    }
-
-    fn is_dirty(&self) -> bool {
-        self.mem_table.is_dirty()
-    }
-
-    #[allow(clippy::unused_async)]
-    async fn init(&mut self, options: InitOptions) -> StorageResult<()> {
-        assert!(
-            self.epoch.replace(options.epoch.curr).is_none(),
-            "epoch in local state store of table id {:?} is init for more than once",
-            self.table_id
-        );
-
-        Ok(())
-    }
-
-    fn seal_current_epoch(&mut self, next_epoch: u64, opts: SealCurrentEpochOptions) {
-        assert!(!self.is_dirty());
-        if let Some(value_checker) = opts.switch_op_consistency_level {
-            self.mem_table.op_consistency_level.update(&value_checker);
-        }
-        let prev_epoch = self
-            .epoch
-            .replace(next_epoch)
-            .expect("should have init epoch before seal the first epoch");
-        assert!(
-            next_epoch > prev_epoch,
-            "new epoch {} should be greater than current epoch: {}",
-            next_epoch,
-            prev_epoch
-        );
-        if let Some((direction, watermarks)) = opts.table_watermarks {
-            let delete_ranges = watermarks
-                .iter()
-                .flat_map(|vnode_watermark| {
-                    let inner_range = match direction {
-                        WatermarkDirection::Ascending => {
-                            (Unbounded, Excluded(vnode_watermark.watermark().clone()))
-                        }
-                        WatermarkDirection::Descending => {
-                            (Excluded(vnode_watermark.watermark().clone()), Unbounded)
-                        }
-                    };
-                    vnode_watermark
-                        .vnode_bitmap()
-                        .iter_vnodes()
-                        .map(move |vnode| {
-                            let (start, end) =
-                                prefixed_range_with_vnode(inner_range.clone(), vnode);
-                            (start.map(|key| key.0.clone()), end.map(|key| key.0.clone()))
-                        })
-                })
-                .collect_vec();
-            if let Err(e) = self.inner.ingest_batch(
-                Vec::new(),
-                delete_ranges,
-                WriteOptions {
-                    epoch: self.epoch(),
-                    table_id: self.table_id,
-                },
-            ) {
-                error!(error = %e.as_report(), "failed to write delete ranges of table watermark");
-            }
-        }
-    }
-
-    async fn try_flush(&mut self) -> StorageResult<()> {
-        Ok(())
-    }
-
-    fn update_vnode_bitmap(&mut self, vnodes: Arc<Bitmap>) -> Arc<Bitmap> {
-        std::mem::replace(&mut self.vnodes, vnodes)
-    }
-
-    fn get_table_watermark(&self, _vnode: VirtualNode) -> Option<Bytes> {
-        // TODO: may store the written table watermark and have a correct implementation
-        None
     }
 }
